week 1/requirements.txt
pandas
numpy
transformers
torch
matplotlib
seaborn
praw
google-api-python-client
scikit-learn
requests
tqdm
python-dotenv
regex
